requirements.txt
# requirements.txt

Flask
python-dotenv
google-cloud-aiplatform
requests
google-auth